% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmethods.R
\name{getmethods}
\alias{getmethods}
\alias{ad_get_table_names}
\alias{ad_get_table_header}
\alias{ad_get_table_data}
\alias{ad_get_table_size}
\alias{ad_get_table}
\title{Get Methods for the ABMI Data API.}
\usage{
ad_get_table_names()

ad_get_table_header(table)

ad_get_table_data(table, site = NULL, year = NULL, skip = 0L, take = 1000L)

ad_get_table_size(table)

ad_get_table(table, site = NULL, year = NULL)
}
\arguments{
\item{table}{Title of a given table, e.g. \code{"T01A"}, es returned by \code{ad_get_table_names()}.}

\item{site}{Site name (return sites contains the given value).}

\item{year}{Year (return records in the given year).}

\item{skip}{Start from which row (default 0).}

\item{take}{Total rows to return (default 1000).}
}
\value{
\code{ad_get_table_names} returns a named character vector with table names and descriptions.

\code{ad_get_table_header} returns a character vector with header names.

\code{ad_get_table_size} returns a single numeric value with the size of the table.

\code{ad_get_table_data} returns a data frame.

\code{ad_get_table} returns a data frame.
}
\description{
Function names are prepended with \code{ad_} referring to ABMI data.
}
\details{
\code{ad_get_table_names} lists names of tables available for download,
\code{ad_get_table_header} lists table headers,
\code{ad_get_table_size} gives the size of the table,
\code{ad_get_table_data} gets a chunk of a table,
\code{ad_get_table} gets the entire table.
}
\examples{
\dontrun{

ad_get_table_names()
ad_get_table_header("T01A")
ad_get_table_data("T01A", take=10)
ad_get_table_size("T01A")
str(ad_get_table("T01A", year=2010))

}

}
\seealso{
\code{\link[=ad_convert_na]{ad_convert_na()}} for dealing with missing value indicators.
}
